+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@
 
 - feat: breaks `readState` into `signed` and `unsigned` methods. The `signed` method will take a pre-signed request and re-use it for polling. The `unsigned` method will create a new request and sign it for each poll.
   - This can be called directly in the `HttpAgent` or through the `Actor` interface, with a new `signed` polling strategy.
+- fix: fixes a bug in the `Principal` library where the management canister id util was incorrectly importing using `fromHex`
 - feat: change auth-client's default identity provider url
 
 ## [2.4.0] - 2025-03-24